add project setup screen domain views
